# Optimize pattern matching